requirements.txt
# Agentic AI Portfolio - Requirements
# Compatible Python version: 3.11
# Install: pip install -r requirements.txt

# ── Streamlit ────────────────────────────────────────────────
streamlit>=1.35.0

# ── CrewAI ──────────────────────────────────────────────────
crewai>=1.9.3
litellm>=1.81.15

# ── MCP ─────────────────────────────────────────────────────
fastmcp==2.3.0
mcp==1.26.0

# ── LangChain stack ─────────────────────────────────────────
langchain==1.2.10
langchain-community==0.4.1
langchain-core==1.2.15
langchain-groq==1.1.2
langchain-huggingface==1.2.0
langchain-mcp-adapters==0.2.1
langchain-tavily==0.2.17
langchain-text-splitters==1.1.1

# ── LangGraph ───────────────────────────────────────────────
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8

# ── HuggingFace stack ───────────────────────────────────────
huggingface-hub==0.36.2
transformers==4.57.6
tokenizers==0.22.2
sentence-transformers==2.7.0
safetensors==0.7.0

# ── Groq ────────────────────────────────────────────────────
groq==0.37.1

# ── Pydantic ────────────────────────────────────────────────
pydantic==2.12.5
pydantic-settings==2.13.1

# ── Vector search ───────────────────────────────────────────
faiss-cpu==1.13.2

# ── PDF loading ─────────────────────────────────────────────
pypdf>=3.0.0

# ── Data / ML ───────────────────────────────────────────────
numpy==2.4.2
scipy==1.17.1
scikit-learn==1.8.0
torch==2.10.0
pandas>=2.1.0

# ── HTTP / Async ────────────────────────────────────────────
httpx==0.28.1
httpx-sse==0.4.3
requests==2.32.5
anyio==4.12.1
aiohttp==3.13.3

# ── ASGI server ─────────────────────────────────────────────
uvicorn==0.41.0
fastapi==0.133.0
sse-starlette==3.2.0

# ── Utilities ───────────────────────────────────────────────
python-dotenv==1.2.1
PyYAML==6.0.3
beautifulsoup4>=4.12.0
tenacity==9.1.4
tqdm==4.67.3
rich==14.3.3
typing_extensions==4.15.0